refactor: Separate SamplesApp.App into multiple partials
chore: Update projitems for SamplesApp

chore: Adjustments
