+++ src/SamplesApp/SamplesApp.Shared/SamplesApp.Shared.projitems
@@ -19,6 +19,8 @@
     <Compile Include="$(MSBuildThisFileDirectory)App.xaml.cs">
       <DependentUpon>App.xaml</DependentUpon>
     </Compile>
+    <Compile Include="$(MSBuildThisFileDirectory)App.Assertions.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)App.Tests.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)App.UnoIslands.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MainPage.xaml.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
